+++ src/Ema/Generate.hs
@@ -10,6 +10,7 @@ import Data.LVar qualified as LVar
 import Data.Some
 import Ema.Asset (Asset (..))
 import Ema.CLI qualified as CLI
+import Ema.Route (PartialIsoEnumerableWithCtx)
 import Ema.Site
 import System.Directory (copyFile, createDirectoryIfMissing, doesDirectoryExist, doesFileExist, doesPathExist)
 import System.FilePath (takeDirectory, (</>))