+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 - Fix bug in `Ex02_Basic.hs` (wasn't generating HTML files)
 - Fix `runEmaPure` not generating routes
-- Typeclass
-  - Change typeclass to be indexed only route (`Ema r`)
+- Multisite
+  - Typeclass is gone, replaced by `Site` datatype that can easily be composed
+  - Main thread can now exit (if no model updates are necessary); no need for `threadDelay maxBound` hacks.
 
 ## 0.6.0.0 -- 2022-02-05